# test: Fix batch predict unit test flakiness